Move pkg/module/model into pkg/module
